Working fixes
- Add point cleaner when using a customPolylineOptions
- Adjust Distance and Duration types
- Add callback management to mainScope

+++ easyroutes/src/main/java/com/tonyakitori/inc/easyroutes/model/Distance.kt
@@ -8,5 +8,5 @@ data class Distance(
 	val text: String? = null,
 
 	@field:SerializedName("value")
-	val value: Int? = null
+	val value: Long? = null
 )
