+++ easyroutes/src/main/java/com/tonyakitori/inc/easyroutes/model/Duration.kt
@@ -8,5 +8,5 @@ data class Duration(
 	val text: String? = null,
 
 	@field:SerializedName("value")
-	val value: Int? = null
+	val value: Long? = null
 )
